Add files via upload
This is a game i tried to code in pygame. I followed a tutorial from a guy and made afew changes to the code i learnt from him and created this game. Its a very simple game where all you need is to dodge some falling balls. The game still needs a lot of fixes and features like scores added and updated to your database or a notepad file and blah blah.
I'm just a newbie so please help me with this. Your feedback will be highly appreciated. Thank you.